creating the modules folder and added some modules
