+++ src/index.html
@@ -40,6 +40,6 @@
         </div>
       </article>
     </section>
-   <script src="main.js"></script>
+   <script type="modules" src="main.js"></script>
   </body>
 </html>
