+++ src/modules/ids.js
@@ -1,0 +1,9 @@
+const nameInpt = document.getElementById('nameInpt');
+const scoreInpt = document.getElementById('scoreInpt');
+const listContainer = document.getElementById('listContainer');
+const form = document.getElementById('form');
+const refreshBtn = document.getElementById('refreshBtn');
+
+export {
+  nameInpt, scoreInpt, listContainer, form, refreshBtn,
+};
