+++ src/modules/printLeaderboard.js
@@ -1,0 +1,12 @@
+import { listContainer } from './ids';
+
+const printUsersAndScores = (data) => {
+  listContainer.innerHTML = '';
+  data.forEach((el) => {
+    listContainer.innerHTML += `
+      <li class="scores__list">${el.user}: ${el.score}</li>
+    `;
+  });
+};
+
+export default printUsersAndScores;
